+++ src/botPage/view/deriv/components/tour/index.js
@@ -1,3 +1,5 @@
 import Tour from './tour.jsx';
+import TourTargets from './tour-targets.jsx';
 
+export { TourTargets };
 export default Tour;
